Restructure note encryption code: move note_encryption.rs functionality to domain.rs, move note_bytes.rs functionality to orchard_domain.rs (and delete note_bytes.rs), and other monor fixes
